requirements.txt
absl-py==2.3.1
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
Automat==25.4.16
babel==2.16.0
bandit==1.8.6
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
certifi==2024.8.30
cffi==2.0.0
charset-normalizer==3.0.1
click==8.1.7
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
distro==1.9.0
Django==5.0.2
django-environ==0.12.0
djangorestframework==3.17.1
dnspython==2.8.0
drozer==3.1.0
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.12.4
Flask==3.0.3
flatbuffers==25.9.23
fonttools==4.59.2
fqdn==1.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.75.0
gunicorn==25.3.0
h11==0.14.0
h5py==3.14.0
httpcore==1.0.5
httpx==0.27.2
hyperlink==21.0.0
idna==3.8
Incremental==24.11.0
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.5.3
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml_dtypes==0.5.3
mysqlclient==2.2.4
namex==0.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.3.2
opt_einsum==3.4.0
optree==0.17.0
overrides==7.7.0
packaging==24.1
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.4
Pillow==10.0.0
platformdirs==3.10.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==6.32.1
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.18.0
PyMySQL==1.1.1
pynput==1.8.1
pyOpenSSL==26.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-whois==0.9.6
pytz==2025.2
pywin32==305
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
reportlab==4.4.4
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.1.0
rpds-py==0.20.0
scikit-learn==1.8.0
scipy==1.16.3
Send2Trash==1.8.3
service-identity==24.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.4.4
stack-data==0.6.3
stevedore==5.5.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
Twisted==25.5.0
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
virtualenv==20.24.5
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
widgetsnbextension==4.0.13
wrapt==1.17.3
xlwt==1.3.0
zope.interface==8.2
